requirements.txt
flask
pandas
numpy
google-genai
python-dotenv
gunicorn
